backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
chromadb==0.4.18
google-generativeai==0.4.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
mysql-connector-python==8.2.0
pydantic==2.12.4
pydantic-settings==2.0.3
email-validator==2.1.0

# LangChain packages
langchain==0.1.20
langchain-google-genai==1.0.1
langchain-chroma==0.1.1
langchain-community==0.0.38
sentence-transformers==3.0.1